chore: extract last function from formula-types to mathjs-utils, add unit tests [PT-186452591]
